# autoevolve-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Company Cockpit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/company/ui/session": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Session Status",
        "description": "Validate dashboard Basic credentials without reading a product domain.",
        "operationId": "session_status_company_ui_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/company/ui/integrations": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Integrations View",
        "operationId": "integrations_view_company_ui_integrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsView"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/company/ui/settings": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Settings View",
        "operationId": "settings_view_company_ui_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IntegrationsView": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/ProviderGroup"
            },
            "type": "array",
            "title": "Groups"
          },
          "total_groups": {
            "type": "integer",
            "title": "Total Groups"
          },
          "groups_with_configuration": {
            "type": "integer",
            "title": "Groups With Configuration"
          },
          "fully_configured_groups": {
            "type": "integer",
            "title": "Fully Configured Groups"
          },
          "configured_keys": {
            "type": "integer",
            "title": "Configured Keys"
          }
        },
        "type": "object",
        "required": [
          "groups",
          "total_groups",
          "groups_with_configuration",
          "fully_configured_groups",
          "configured_keys"
        ],
        "title": "IntegrationsView"
      },
...
      "ProviderGroup": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "configured_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Configured Keys"
          },
          "missing_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Keys"
          },
          "configured_count": {
            "type": "integer",
            "title": "Configured Count"
          },
          "total_keys": {
            "type": "integer",
            "title": "Total Keys"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "category",
          "capabilities",
          "configured_keys",
          "missing_keys",
          "configured_count",
          "total_keys"
        ],
        "title": "ProviderGroup"
      },
...
      "PublicLink": {
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "app",
              "forms",
              "calendar"
            ],
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "state": {
            "type": "string",
            "enum": [
              "local",
              "placeholder",
              "configured",
              "invalid"
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "url",
          "state"
        ],
        "title": "PublicLink"
      },
...
      "SessionStatus": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated"
          }
        },
        "type": "object",
        "required": [
          "authenticated"
        ],
        "title": "SessionStatus"
      },
      "SettingsView": {
        "properties": {
          "orgs": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceOrg"
            },
            "type": "array",
            "title": "Orgs"
          },
          "active_org_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Org Id"
          },
          "available_capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Capabilities"
          },
          "public_links": {
            "items": {
              "$ref": "#/components/schemas/PublicLink"
            },
            "type": "array",
            "title": "Public Links"
          }
        },
        "type": "object",
        "required": [
          "orgs",
          "active_org_id",
          "available_capabilities",
          "public_links"
        ],
        "title": "SettingsView"
      },
...
      "WorkspaceOrg": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Capabilities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "domain",
          "capabilities"
        ],
        "title": "WorkspaceOrg"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}